[REFACTOR/#52] 학생 제휴 건의함 페이지 리팩토링
