+++ src/pages/student/suggestion-form/api/index.ts
@@ -1,0 +1,2 @@
+export type { PostSuggestionPayload } from "./postSuggestion";
+export { postSuggestion } from "./postSuggestion";
